Add files via upload
CAD work- integrating motors and gears into the frame to mechanize it.

The first build is going to be small, about 1 foot across, and use $3 stepper motors.

This will allow me to design and build it far faster and cheaper than the large version. Then I can focus on the software controlling the framework sooner than if I went slowly and instead built the large frame first.